Use XML-Scanner in game-core to do shallow-parsing
The shallow-parsing is so we can find the game name of a XML file
without doing the full parse of XML and without assembling
a GameData object.

This update changes shallow parsing to use a new XML Scanner
(in xml-reader) that is based on StAX.

The performance change for Bot startup speed is negligable, but
for listing game files (clicking 'select map' button), there
is a pretty nice speedup and XML shallow parsing time is
reduced from on average 80ms (per XML) to 0.5ms

Forward-looking, future work will incorporate the XML body-text
scanner to read game-notes in a quick manner to speed up the listing
of game notes in the 'select game' screen.