Add map-data Java Model (add map-data subproject with POJOs)
This update adds model objects that can work with the 'xml-reader' library
and be instantiated via the XmlMapper (in 'xml-reader').  The model
objects are meant to strictly represent what is in the XML files without
any further interpretation of the data. Once these objects are loaded
from an XML we'll then need to go through an "assembly" process (future
work) to convert these objects into a 'GameData'.