Update: Improve map parsing time - more efficient player attachment lookup (#7585)
* Update: Improve map parsing time - more efficient player attachment lookup

This update improves game parsing speed by increasing the efficiency of
attachment processing. This update converts the attachment processing algorithm,
used on full-game-parse, from a runtime of O(p * a), where p is number of players
and 'a' is the number of attachments per player to O(p).

The core of this update is taking advantage that we map attachments by their
name and we can do a O(1) lookup by name. The previous algorithm iterated over
every attachment looking an attachment with a given name.

For full map parsing, here are some example performance numbers before
and after this update:
- War of the Relics: 4751 ms -> 1569.8 ms
- World at war:      3200 ms -> 1102.7 ms
- World of Warcraft: 2783 ms -> 1147.1 ms

* Update test case, trigger attachments should be added with their name as key

* Update: Improve further usages of 'getTriggers'

- Use non-null argument for predicate match of getTrigger
- Again use the player attachment get method rather than
searcing for attachments within all of a players attachments

Co-authored-by: tripleabuilderbot <tripleabuilderbot@gmail.com>