Fix loading of deleted maps (#7611)
If you load a map and it then becomes the 'default game', if you
then close the game and delete the map and then launch the game,
there are errors trying to load the 'default game'. This update
fixes that such that if a default game does not exist on disk
then we do not try to load it.