Add a 'setup' ansible group & move ops docs to '/infrastructure'
The idea is we will have a special group that new boxes
can *also* be added to for an initial setup to be run.
Once boxes receive the setup, we can move them out
of that group to avoid redundnat operations being run on
them.

Note, we keep admin_users as part of the common deployment
run every time for a few reasons:
- if we add or change an admin user (add a SSH key),
  we would not want to migrate all machines into the setup
  group and then back out
- the role execution is very reliable and quick