Refactor roll dice, select casualties, and remove casualties steps
The Fire and FireAa classes are no longer used in building the 3 fire
sub steps.

Each of the main fire steps (OffensiveAaFire, OffensiveGeneral,
OffensiveFirstStrike, etc) build FiringGroups using a FiringGroupFilter.
These FiringGroups are used by FireStepsBuilder to generate a set of
fire sub steps for each group. The groups have their own unique name and
is displayed in the battle ui.

The AirVsNonSubs step is removed because that firing group is now named
and will be displayed in the battle ui when it fires, selects, and
removes.