Re-add warning about save from a newer compatible version (#7871)
* Re-add warning about save from a newer compatible version

This was removed in 030cb1f61d574a2d87fc4eb6777f953fec642496 because
of a swing deadlock issue and was also considered to be redundant.

It is needed because newer compatible versions might use new saved data
to add or change functionality. When an older version opens the saved
game, the new saved data will be removed and the functionality might end
up broken.