Make sure the default game is there or clear it if it isn't
If the default game is not on the system, it wasn't being cleared out of
the game.  This would leave the game in an odd state where it is
partially loaded but has unused objects listening for events.

By calling resetDefaultGame when the default game is missing, the game
can be reset to an initial state where no map has been selected.

In addition, the checks to see if the file exists had issues when the
uri had special prefixes and suffixes.  So those are stripped before the
checks.