Remove need for infrastructure setup step (#7930)
This update removes the setup step by adding conditional checks
that skip the core UFW setup commands if they are not needed.

Deployments were previously broken up into two phase, a one time
setup and then a regular deployment. A problem with this is
that it was two step and required updates to remove machines
from the setup inventory group.

The reason for the split was to avoid re-running UFW commands
which have a tendency to fail or collide.