Fix AA support calculation
All of the friendly and enemy units need to be used to calculate the
support during casualty selection.

When getting the total power and total strength values for the
defendingAa, the "defending" value needs to be flipped since the
"defending" value is the side of the targeted planes and not the side
of the AA.  And the enemy/friendly units need to be flipped as well.