Stateful injections (#8039)
* Convert 'Injections.java' to be stateful

* Init Injections in top level classes, pushes init out of game-core into top level sub-projects

This is key to allow for the top-level dependencies to declare their own implementations and
inject them into the Injections. For example, a game-headed specific implementation can now
be defined in that sub-project, not visible to game-core nor headless-game, and be injected
into injections.

* Encapsulate Injections.instance so it can only be initialized via 'init' method

This enforces that injections are only initialized once (write-once, then read-only)

* Encapsulate Injections data access with getters

This allows for a more uniform access to data in injections and easier construction.

* Update Javadoc on Injections.java

Describe that Injections properties should be constructor injected and not used in a static context.