Fix, prevent NPE in FileSystems#getFileSystem when default game URI is missing a scheme (#8058)
Bug report #8015 has a NPE in getFileSystem, this looks to happen if the default
game URI is missing a scheme. That can be something trivial like an empty
value for the default game URI (empty but not null), or could be caused by
other reasons.

Regardless, instead of throwing an exception in such cases, we can fall back to resetting
the default game to a default value.