Allow power to go over diceSides in AirBattle
If an air unit had 2 rolls with a strength equal to the dice sides, that
unit should hit 2 targets in low luck. But with the power limited to
dice sides, the unit would only be able to hit 1 target. The dice sides
limit is not used in any other power calculation, either.