Break up 'http-clients' subproject
This update breaks up http-clients and creates a number
of new subprojects:
- github-client
- websocket-client
- lobby-client
- maps-server-client

http-clients now becomes a generic base-subproject for
other http-client specific subprojects (EG: lobby-client)

Goal is to move java files from http-clients to
more specific subprojects and allow for 'websocket-client'
specific code to be in its own subproject rather than
mixed with http-clients. This partitioning allows for
dependencies on just websoket-client without bringing
in the rest of http-clients.

Secondarily, the http-clients are useful to split up
themselves as the clients can be quite different and
benefit from partitioning.