Unify low luck and normal roll code in DiceRoll
The AA low luck and normal rolls have the same logic as the non-AA low
luck and normal rolls. The only difference is handling dice sides that
are different from the game state.  So add a getDiceSides method to
TotalPowerAndTotalRolls and unify the code for low luck and normal
rolls.