Add spotlessCheck to auto-format task (#8237)
This update migrates the formatting check to github workflow.
We add this to the auto-format task to keep the spotless formatting
checks together. Not all spotless violations can be fixed by
running spotlessApply, so we can have cases where spotlessApply
does what it can, results in no-changes, but we still have issues
that would fail spotlessCheck.