Allow validating canals when finding neighbors or distances
The ProAI calculates the value of territories based on neighboring
territories and distances to enemy factories/capitals. If a neighboring
territory had a canal to it, the ProAI would ignore the canal and assume
that the units could travel to it. This could cause unit stacking
because of the higher value of the territory.

Now, when the ProAI calculates the neighboring territories, it checks if
units can travel between between the two territories. This will cause
some neighboring territories to have a higher distance and so the value
of the territory will go down.