Project#20 - Add ClientNetworkBridge to IDelegateBridge, enable websocket send and receive API
- We add a new interface ClientNetworkBridge that will be the main interface for sending
websocket network messages. We can also use this to register listeners to listen for
network messages and invoke the existing network APIs.

The key design difference will be after receving a network message we will know, based
on message type, which method to invoke and the message will contain the method arguments.
This would replace the way methods are invoked previously via remote method calls.

This update only adds the needed API and wires the ClientNetworkBridge to where it will
be needed, but it is not yet used (future commits will build on this)