Rename ChangeAttachmentChange and use primitives
ChangeAttachmentChange was renamed to AttachmentPropertyChange.

ChangeAttachmentChange stores an Attachable object. This is an
interface and didn't have an easy way to turn it into a primitive. A new
method getPrimitiveForm has been added that will generate an
PrimitiveNamedAttachable. The PrimitiveNamedAttachable can be easily
serialized and can deserialize with the help of the GameData object.

The PrimitiveNamedAttachable is based off of GameObjectStreamData.