Remove MainFrame 'exit on close' window closed listener
The exit-on-close assumes there will be no other windows open,
if we are to allow multiple game instances without spawning a new
process we need to avoid a system.exit that would close all windows.
In specific, if we do not spawn a new process for games then there
will be multiple game windows, closing one will render a main frame
which when closed would then invoke a system.exit

Furthermore, the exit-on-close is unnecessary as that preempts
the window closed action that will System.exit if there are no
longer anymore visible windows.