Simplify the Individual and Random AA casualty selection (#8328)
The logic for both the Individual and Random AA casualty selection is
practically identical.  The only difference is when the number of rolls
is equal to the number of targets. And that difference is just in the
fact that a random generator isn't called.

This combines the two methods but keeps the small difference still
around.