Move BattleActions#removeCasualties to BattleState#markCasualties
Remove the code that depends on the bridge to the caller so that
markCasualties can be in the BattleState. The code that depends on the
bridge is old code that will be removed once save compatibility can be
broken.

Modify MustFightBattle#remove to remove the units from the WaitingToDie
lists so that markCasualties can always move the units to the
WaitingToDie list but the caller can then also remove those units if
needed.