Remove the call to remove from damagedChangeInto
MustFightBattle#remove does more work than just remove the unit from the
battle. If the unit type is marked to convert to another unit on death,
it will convert it. And it checks for dependent units and kills them.

Neither of those are needed when converting a damaged unit into another
unit, especially the extra conversion check. If a unit type had a
conversion requested when it had 1 hit point left and 0 hit points left,
it would never do the 1 hit point left conversion because of the extra
conversion in remove().