Create HistoryChange objects for killing and transforming units
The HistoryChange objects will group the calls to add a history child
event and calls to add change objects. This will creating these objects
without the need for the bridge and then returning them to callers where
the bridge is available.

KillUnit object has been used in many of the places in the battle code
that removes units. In most of those places, transformation wasn't being
done, so this fixes that.

KillUnit also makes it explicit that transported and unloaded units are
killed when it is killed. There is still some code (such as
AbstractBattle#unitsLostInPrecedingBattle) which does some of this work,
though.

Rework how units are killed and transformed in MustFightBattle. The
original logic went like:
1. Create a change object to kill a unit
2. Detect if the unit needs to transform when it dies
3. Create a change object to also set its damage to the maximum
 - At this point, there is a change object to kill it and one to set its
 damage
4. Add the units to a list of killed units during the round
5. Execute the change objects
6. Later, during the clearCasualty steps, go through the units hit as
 well as the units that were previously killed in step 4
7. Transform them

If the change object in 3 doesn't happen, then the unit won't transform
because its damage won't match the settings.
If the unit isn't added to the list in 4, then the unit won't transform
because it won't be seen in step 6.

The new logic now has the code to kill units moved to the KillUnits
HistoryChange object. This object will call TransformUnits if needed.
So the new logic looks like this:
1. Create KillUnits object
2. This creates change objects to kill the units
3. This creates a TransformUnits object that creates change objects to
 transform the units.
4. Later, during the clearCasualty step, only the units that have been
hit get passed into a TransformUnits object. The killed units from 1-3
no longer need to be worried about.

Because of save compatibility reasons, the old killedDuringCurrentRound
variable is still around and is still being used in step 4. But it is no
longer being populated so it will only actually be used if the save is
in the middle of a battle and has already populated it.