Remove request & response logging feature from lobby server
The feature depends on java.util.logging which we are moving
away from. The feature is toggled by a debug flag that is
currently off, and has not been used in some time.

For the above reasons, this update removes the request & response
logging. If we need it again as a debug tool we can build it
back in.