Fix concurrent modification exception when setting player colors
When launching the game, we can get a 'ConcurrentModificationException'
when player colors are assigned.

This fix reverts to using a non-atomic add to the player color map
and removes usage of 'computeIfAbsent' (which is not thread-safe).

Because we set the player color to the same value, but possibly doing
so concurrently, we do not need a thread-safe nor atomic add of the
player color (we just need it to work). By going to a non-atomic
version, we avoid code that would otherwise throw a
'ConcurrentModificationException' on concurrent adds.