Initial work on how units target other units
Currently, a battle assumes that it always has 4 firing phases: AA,
Bombardment, First Strike, and General/Other. At the start of each of
the phases, it uses one of the "FiringGroupSplitter" classes to group
the firing units and their targets into "FiringGroups". It then loops
through all of the "FiringGroups" and creates a set of "RollDiceStep",
"SelectCasualties", and "MarkCasualties" battle steps.

The new logic is to create a list of "BattleGroups" at either the very
beginning of a battle or at the beginning of the BattleDelegate. The
list of BattleGroups is constant for all battles and so it only needs to
be built once.

Each BattleGroup has 1 or more FiringSquadrons. These FiringSquadrons
have predicates that are used to determine what units are firing in it,
what units are targeted in it, and if it should even fire.

In a battle, each BattleGroup will be iterated over and the
active FiringSquadrons will be determined. Then, a set of  battle steps
will be created for each of the FiringSquadrons and the game will
proceed.

This commit only has the work to create the BattleGroups and
FiringSquadrons. They are not being used yet.

This rework will almost replace everything in the
games.strategy.triplea.delegate.battle.steps.fire package.

And end goal is to allow map makers to create their own BattleGroups and
FiringSquadrons. If the map maker creates their own, then the
BattleGroups.createBattleGroups would not be called. It should only be
called for maps that don't have custom BattleGroups/FiringSquadrons.