Remove available games cache
Removes cache of the list of available games. The cache is being
removed in preparation for the available game list to move
to "index" YML files that can be read instead of parsing the
available maps on disk. The index files can be read very quickly
which will make a cache unnecessary.