Pass in UnitTypeList in UnitSupportAttachment.get
The UnitTypeList is a read-only data structure. So the lock in
UnitSupportAttachment.get is unneeded.

With the lock removed, UnitSupportAttachment.get can be changed to just
accept the UnitTypeList instead of GameData.

Added comments to UnitTypeList to indicate that it is read-only.  Added
comments to GameData.getUnitTypeList to indicate that the lock isn't
needed for it.  Added comments to UnitSupportAttachment to indicate that
the set of UnitSupportAttachment is immutable.