Update map versions for maps that have been updated to have relative links
The engine has logic to find the filename of an image link and to
make that filename relative. This is overly complex as the link
could just be relative to begin with.

Various maps have been updated to remove the ignored preceding part
of an image link to just be relative, eg:
{http://sourceforge/path/img.png -> img.png}