Simplify link localization code
Main goal of this simplification is to remove 'map name' from resource loader.
One place that is used is in Link Localization. Instead of passing a resource
loader to the link localizer, we can instead pass a file path which in turn
simplifies how link localization can be done.

Additional:
- removed link localization cache
- changed link localizatoin behavior to only localize links that are not already
  absolute. Previously links in single quotes would be localized and links in
  double quotes would be left as-is. The localization of fully qualified links
  would parse the link path, get the file name and localize the file name path.
  This is removed as it is difficult to understand and is much easier if
  we localize any relative links but keep absolute links as they are (regardless
  of quoting).