Simplify out of date map version lookup
Previously we would compute the mapping of installed maps to versions,
available maps to versions, then compute the union of these two maps
and return any maps that had an installed version less than what
was available.

This update changes that algorithm to instead to iterate over the
list of available maps, find that map, find the version for that
map, and if out of date the map is out of date.

Of note, rather than computing a Map of available map names to
versions, we directly use 'DownloadFileDescription'.