Add flag to bot download-map script to update maps
Adds a boolean flag to the download-map script where if
the flag is true we will always download any given map.
Previously we would download maps on bots only if they
did not already exist, with this flag we can update maps.

A helper script is added to run this operation on bots:
'run_ansible_update_bots'