Fix error prone warning TypeParameterUnusedInFormals
> YamlReader.java:76: warning: [TypeParameterUnusedInFormals] Declaring a type parameter that is only used in the return type is a misuse of generics: operations on the type parameter are unchecked, it hides unsafe casts at invocations of the method, and it interacts badly with method overload resolution.
  private static <T> T readYaml(final InputStream inputStream) {
                       ^

The warning is accurate as class cast exception was thrown at
the caller level and for this reason this method was made private
and the value cast by caller methods.

To resolve this warning, the return type is now an explicit 'Object'
and the caller methods do an explicit type cast.