Fix banned username matching
The check for username bans attempted to do a case insensitive match
by lower casing the name to check. But, the names stored where in whatever
casing was originally entered. This implies the username ban only worked
when the original casing was all lower casing.

This update fixes the matching so it is case insensitive, banned usernames
are upper-cased when entered into database and then the name to match is also upper-cased.