Simplify create-release github action & use unique tags
Deleting and recreating tags does not consistently replace a release.
We are observing the replacement of the tag but the release can then
be created as a draft. This is not what we want.

This update simplifies the steps in the create-release action
and instead of deleting and re-using the 'prerelease' tag,
instead the tag will just match the product version
and will be unique each time.