Improve failed to load 'polygons.txt' error message
If the resource loader path is bad and cannot locate the 'polygons.txt'
file witin a map, an error message is displayed that is both very
wordy and does not give much information about the problem.

This update changes the error message to list the ResourceLoader
search paths and reduces the wordiness of the error message.