Update all preview images to be 'preview.png'
After having downloaded all preview images and uploaded them to each
map repository as 'preview.png' in the top-most folder, this update
changes triplea_map.yaml 'img' attribute to point the preview
image to this file.

With a standard file name we will be able to have the game-client
download the preview.png directly if it knows the URI of the map repo.
This also puts us in a place where map makers can define their preview
image and bundle it with their map.