Fix download file unzip folder move
Overview:
- fixes downloaded map extraction folder name
- cleans up FileUtils

Problem: we download to a temp file and it is the temp file that is moved
(without rename) to the downloaded-maps folder.

Fix: Instead of downloading to an arbitrary temp file, we instead create
a temp folder and download to a file that has the target file name.
We then extract that into downloadedMaps with the correct folder name.

This problem was introduced when handling the case of variants being extracted
from downloadedMaps folder and expanded in place.