Add Map server indexing scheduled job
Adds a scheduled job to the maps-server to run periodically and scan
all map repositories for a 'map.yml' file, index it, and add the
indexed information to database. That DB information is then used
later to server map-download requests to get the list of maps
available for download.

Overview:
- add scheduled indexing job
- add API client to query github orgs for repo listing
- add maps-server config to inject github credentials
- updates maps-server DB, recreates tables to fit necessary information.
  Description and preview image will be at a standard location, we will
  not need to store in DB, game clients can determine those locations
  dynamically based on the map repo URL