Update bot deployment - always deploy zip file
Instead of deploying a jar file for 'latest', instead
run 'releases' task to generate a zip file and copy that.
When not using latest, we download the zip file, by also
using a zip file in the latest case, we get similar steps
where the zip file after being copied or downloaded is
then always extracted.