Restore explicit CLI update of DB password and align passwords
The ansible postgres_user role seems to _not_ be updating passwords.
On deployment am seeing invalid password. This update restores
the CLI 'hack' of setting the database user passwords explicitly.

Second, passwords are updated and aligned for default cases to
be 'lobby' and 'maps' for respecively the 'lobby_db' and 'maps_db'.
That update is so we are not sharing the same password with the
postgres user, which gives us more confidence we are using the right
DB account when accessing database.