Fix flyway to actually run migrations
Flyway was silently failing after running 0 migrations. This update changes
the classpath search arg to a file system search, allows flyway to find
the migrations (and also regrettably so far requires flyway to be
run as root as well to find thes files).

Also, improve error detection of flyway task to mark task as failed
if it exits in error or warns about not finding any migrations.