Always sort unit categories
The units categories were intentionally not sorted in the BattleDisplay
because it needed the categories sorted by strength so that it could do
support calculations.  But the logic for support calculations has
changed to the point that the categories don't need to be sorted by
strength. So categories can be sorted without causing a problem.