Update Oil and Snow with Name Change
So i tried removing 1st edition everywhere in the existing map folders an xml and it would show up with the description picture but just go blank when I selected it. I noticed where it was originally forked from it's named Oil and Snow without the 1st Edition. 

At any rate, if the yaml mapName here is the same as the triplea-maps repo one  and it automatically adds underscore to spaces and makes it all lower case, then I think we should be fine. I don't really know how to test until it's added though. I did rename the zip in downLoaded maps with 1st_edition so I think it should work.

Just found it here. Was looking at the bottom of the list for it lol :)