Switch back to 'flow' style for map.yml generation
The flow structure would have all elements aligning on
the left hand edge of a map.yml file. While this could
be more difficult to diff and automatically update, it
does make the indentation concerns easier to deal with.
It has been observed that getting indentation right
has been a stumbling block for map makers and YML files.

For example, instead of a map being generated as:
- this_is_map_key: value
  key2: valu2

We will now generate it as:
- {this_is_map_key: value, key2, value2}