Validate folder path overrides
- Validate folder path override before setting it, if set
to an invalid value (DNE or cannot be written), then show a warning
a message and reject the value
- Secondly, if a map folder override value becomes invalid, show
a warning value to user and reset the override value. This can
happen for example if the map folder is locked or deleted and can
(for some reason) not be created.

Fixes: https://github.com/triplea-game/triplea/issues/5485