Replace 'none' territory selected with an empty string
In the status bar we print the name of the territory as
the mouse cursor changes. When hovering over a territory
boundary the territory text changes to 'none'. Instead
of this behavior, this update prints an empty string
instead of the text 'none'. So we display nothing
rather than the text 'none'.