Fix NPE when using 'ctrl-a' to add attackers with no territory selected
This updates prevent a NPE from occuring when no territory is
currently selected (hover over a territory border) and then pressing
'ctrl+a' to add attackers.

To repro NPE, open battle calculator, hover over a territory border
such that no territory is current, then use 'ctrl+a' to add attackers
and observe a NPE.

The 'add defenders' feature already this NPE fix, the add attackers
is missing it.