Remove dev-override file
Dev override file was intentioned to be an easier way to create
dev specific settings without using the client settings framework.
This did not receive general usage and introduces two ways to
create dev overrides which is undesirable.

This is being removed as the plan is to have one http-server
going forward which means the existing dev override mechanisms
in client settings will be sufficient